+++ tests/e2e/concurrent/health.test.js
@@ -1,5 +1,5 @@
 const { it } = require("@jest/globals");
-const { AnythinkClient } = require("./anytinkClient");
+const { AnythinkClient } = require("../anytinkClient");
 
 let anythinkClient;
