+++ tests/e2e/concurrent/profiles.test.js
@@ -1,6 +1,6 @@
 const { test, beforeAll } = require("@jest/globals");
-const { AnythinkClient } = require("./anytinkClient");
-const { randomUserInfo } = require("./utils");
+const { AnythinkClient } = require("../anytinkClient");
+const { randomUserInfo } = require("../utils");
 
 let anythinkClient;
