+++ tests/e2e/jest.concurrent.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  testPathIgnorePatterns: ["/node_modules/", "/public/"],
+  testPathIgnorePatterns: ["/node_modules/", "/sequential/"],
   globalSetup: "<rootDir>/setup.js",
   globalTeardown: "<rootDir>/teardown.js",
 };
