+++ tests/e2e/jest.sequential.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  testPathIgnorePatterns: ["/node_modules/", "/concurrent/"],
+  globalSetup: "<rootDir>/setup.js",
+  globalTeardown: "<rootDir>/teardown.js",
+};
