+++ tests/e2e/package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "test": "yarn jest -c jest.e2e.config.js --maxWorkers=1",
+    "test": "yarn jest -c jest.sequential.config.js --maxWorkers=1 && yarn jest -c jest.concurrent.config.js --maxWorkers=100%",
     "format": "yarn prettier --write .",
     "lint": "yarn eslint . && yarn prettier -c ."
   },
