--- tests/e2e/ping.test.js
@@ -1,17 +1,0 @@
-const { it } = require("@jest/globals");
-const { AnythinkClient } = require("./anytinkClient");
-const { execAndWaitForEvent } = require("./wilcoEngine/utils");
-
-let anythinkClient;
-
-beforeEach(async () => {
-  anythinkClient = new AnythinkClient();
-});
-
-describe("Ping Route", () => {
-  it("Sends ping event to the Wilco Engine", async () => {
-    await execAndWaitForEvent("ping", async () => {
-      await anythinkClient.ping();
-    });
-  });
-});
